+++ pingouin/utils.py
@@ -336,8 +336,7 @@ def _check_eftype(eftype):
         return False
 
 
-def _check_dataframe(data=None, dv=None, between=None,
-                     within=None, subject=None, effects=None):
+def _check_dataframe(data=None, dv=None, between=None, within=None, subject=None, effects=None):
     """Checks whether data is a dataframe or can be converted to a dataframe.
     If successful, a dataframe is returned. If not successful, a ValueError is
     raised.
@@ -347,21 +346,20 @@ def _check_dataframe(data=None, dv=None, between=None,
         # DataMatrix objects can be safely convert to DataFrame objects. By
         # first checking the name of the class, we avoid having to actually
         # import DataMatrix unless it is necessary.
-        if data.__class__.__name__ == 'DataMatrix':
+        if data.__class__.__name__ == "DataMatrix":  # noqa
             try:
-                from datamatrix import DataMatrix, convert as cnv
+                from datamatrix import DataMatrix, convert as cnv  # noqa
             except ImportError:
                 raise ValueError(
-                    "Failed to convert object to pandas dataframe (DataMatrix not available)")
+                    "Failed to convert object to pandas dataframe (DataMatrix not available)"  # noqa
+                )
             else:
                 if isinstance(data, DataMatrix):
                     data = cnv.to_pandas(data)
                 else:
-                    raise ValueError(
-                        "Data must be a pandas dataframe or compatible object.")
+                    raise ValueError("Data must be a pandas dataframe or compatible object.")
         else:
-            raise ValueError(
-                "Data must be a pandas dataframe or compatible object.")
+            raise ValueError("Data must be a pandas dataframe or compatible object.")
     # Check that both dv and data are provided.
     if any(v is None for v in [dv, data]):
         raise ValueError("DV and data must be specified")
@@ -384,8 +382,7 @@ def _check_dataframe(data=None, dv=None, between=None,
     if effects == "interaction":
         for input in [within, between]:
             if not isinstance(input, (str, int, list)):
-                raise ValueError(
-                    "within and between must be specified when effects=interaction")
+                raise ValueError("within and between must be specified when effects=interaction")
     return data